pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "deep_learning_land_use_classification"
version = "0.0.1"
description = "School project to classify land use images"
authors = [
  { name = "Tomer Peled" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "loguru",
    "mkdocs",
    "pip",
    "python-dotenv",
    "ruff",
    "tqdm",
    "typer",
    "matplotlib",
    "pandas",
    "numpy",
    "torchvision",
    "scikit-learn",
    "wandb",
    "nbformat",
    "transformers",
    "iterative-stratification",
    "seaborn"
]
requires-python = "~=3.10.0"


[tool.ruff]
line-length = 99
src = ["deep_learning_land_use_classification"]
include = ["pyproject.toml", "deep_learning_land_use_classification/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["deep_learning_land_use_classification"]
force-sort-within-sections = true

